more fun reshuffles
