+++ internal/command/mpg/restore.go
@@ -1,4 +1,4 @@
-package cmdv1
+package mpg
 
 import (
 	"context"
@@ -7,12 +7,13 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/superfly/flyctl/internal/command"
 	"github.com/superfly/flyctl/internal/command/mpg/utils"
+	cmdv1 "github.com/superfly/flyctl/internal/command/mpg/v1"
+	cmdv2 "github.com/superfly/flyctl/internal/command/mpg/v2"
 	"github.com/superfly/flyctl/internal/flag"
-	mpgv1 "github.com/superfly/flyctl/internal/uiex/mpg/v1"
 	"github.com/superfly/flyctl/iostreams"
 )
 
-func NewRestore() *cobra.Command {
+func newRestore() *cobra.Command {
 	const (
 		long  = `Restore a Managed Postgres cluster from a backup.`
 		short = "Restore MPG cluster from backup."
@@ -37,16 +38,16 @@ func NewRestore() *cobra.Command {
 
 func runRestore(ctx context.Context) error {
 	out := iostreams.FromContext(ctx).Out
-	mpgClient := mpgv1.ClientFromContext(ctx)
+
+	var cluster *utils.ManagedCluster
+	var err error
 
 	clusterID := flag.FirstArg(ctx)
 	if clusterID == "" {
-		cluster, _, err := utils.ClusterFromArgOrSelect(ctx, clusterID, "")
+		cluster, _, err = utils.ClusterFromArgOrSelect(ctx, clusterID, "")
 		if err != nil {
 			return err
 		}
-
-		clusterID = cluster.Id
 	}
 
 	backupID := flag.GetString(ctx, "backup-id")
@@ -56,18 +57,8 @@ func runRestore(ctx context.Context) error {
 
 	fmt.Fprintf(out, "Restoring cluster %s from backup %s...\n", clusterID, backupID)
 
-	input := mpgv1.RestoreManagedClusterBackupInput{
-		BackupId: backupID,
+	if cluster.Version == utils.V1 {
+		return cmdv1.RunRestore(ctx, clusterID, backupID)
 	}
-
-	response, err := mpgClient.RestoreManagedClusterBackup(ctx, clusterID, input)
-	if err != nil {
-		return fmt.Errorf("failed to restore backup: %w", err)
-	}
-
-	fmt.Fprintf(out, "Restore initiated successfully!\n")
-	fmt.Fprintf(out, "  Cluster ID: %s\n", response.Data.Id)
-	fmt.Fprintf(out, "  Cluster Name: %s\n", response.Data.Name)
-
-	return nil
+	return cmdv2.RunRestore(ctx, clusterID, backupID)
 }
