+++ internal/command/mpg/v2/regions.go
@@ -1,4 +1,4 @@
-package cmdv1
+package cmdv2
 
 import (
 	"context"
