+++ internal/command/mpg/v2/run_attach.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunAttach(ctx context.Context) error {
+	return nil
+}
