+++ internal/command/mpg/v2/run_backup.go
@@ -1,0 +1,13 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunBackupList(ctx context.Context, clusterID string) error {
+	return nil
+}
+
+func RunBackupCreate(ctx context.Context, clusterID string) error {
+	return nil
+}
