+++ internal/command/mpg/v2/run_connect.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunConnect(ctx context.Context, clusterID, orgSlug string, proxyPort string) (err error) {
+	return nil
+}
