+++ internal/command/mpg/v2/run_databases.go
@@ -1,0 +1,13 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunDatabasesList(ctx context.Context, clusterID string) error {
+	return nil
+}
+
+func RunDatabasesCreate(ctx context.Context, clusterID string) error {
+	return nil
+}
