+++ internal/command/mpg/v2/run_destroy.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunDestroy(ctx context.Context, clusterID string) error {
+	return nil
+}
