+++ internal/command/mpg/v2/run_detach.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunDetach(ctx context.Context, clusterID string, appName string) error {
+	return nil
+}
