+++ internal/command/mpg/v2/run_proxy.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunProxy(ctx context.Context, clusterID string, proxyPort string, orgSlug string) error {
+	return nil
+}
