+++ internal/command/mpg/v2/run_restore.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunRestore(ctx context.Context, clusterID string, backupID string) error {
+	return nil
+}
