+++ internal/command/mpg/v2/run_status.go
@@ -1,0 +1,9 @@
+package cmdv2
+
+import (
+	"context"
+)
+
+func RunStatus(ctx context.Context, clusterID string) error {
+	return nil
+}
